refactor: rename projects utils folder
