+++ packages/forms-web-app/src/pages/project-search/utils/get-page-data.js
@@ -3,8 +3,8 @@ const { buildPaginationQueryString } = require('../../_utils/build-pagination-qu
 const { mapApplications } = require('../../_utils/map-applications');
 const { mapTitles } = require('../../_utils/map-titles');
 const { allProjectsSubNavigationRoutes } = require('../../_utils/sub-navigation-routes');
-const { documentsPerPage } = require('../../projects/utils/pagination/documentsPerPage');
-const { getPagination } = require('../../projects/utils/pagination/pagination');
+const { documentsPerPage } = require('../../projects/_utils/pagination/documentsPerPage');
+const { getPagination } = require('../../projects/_utils/pagination/pagination');
 const { getFilters } = require('./filters/get-filters');
 const { getProjectSearchSortByLinks } = require('./get-project-search-sort-by-links');
