+++ packages/forms-web-app/src/pages/projects/project-updates/utils/get-formatted-updates.js
@@ -1,4 +1,4 @@
-const { formatProjectUpdate } = require('../../utils/format-project-update');
+const { formatProjectUpdate } = require('../../_utils/format-project-update');
 
 const getFormattedUpdates = (updates) => updates.map((update) => formatProjectUpdate(update));
